requirements.txt
numpy==1.18.1
pandas==1.0.1
pyodbc==4.0.28
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
azure-core==1.7.0
azure-identity==1.3.1
azure-keyvault-secrets==4.1.0
python-dotenv==0.14.0
